+++ lib/jazzy/config.rb
@@ -22,6 +22,7 @@ module Jazzy
     attr_accessor :root_url
     attr_accessor :version
     attr_accessor :min_acl
+    attr_accessor :document_all
 
     def initialize
       self.output = Pathname('docs')
@@ -33,6 +34,7 @@ module Jazzy
       self.docset_platform = 'jazzy'
       self.version = '1.0'
       self.min_acl = SourceDeclaration::AccessControlLevel.internal
+      self.document_all = false
     end
 
     # rubocop:disable Metrics/MethodLength
@@ -119,6 +121,12 @@ module Jazzy
           elsif acl == 'public'
             config.min_acl = SourceDeclaration::AccessControlLevel.public
           end
+        end
+
+        opt.on('--[no-]document-all',
+               'Generate documentation for all declarations, not just ones \
+                with documentation comments.') do |document_all|
+          config.document_all = document_all
         end
 
         opt.on('-v', '--version', 'Print version number') do